Add files via upload
This Matlab code first calculates the scattering, extinction and absorption cross-section of a spherical particle and lets you choose at which energy you want to calculate the scattered and internal fields. Then it calculates the scattered and internal fields in a 3D box at the chosen energy using the formalism in Bohren&Huffman. In agreement with the book's system of reference, the incident plane wave propagates along the z direction, with the electric field polarized along the x-axis. The
% polar angle 'theta', defined as the angle between the scattered vector and the z-axis, spans from 0 degrees (forward scattering) to 180 degrees (back scattering), the azimuthal angle 'phi', defined as the angle between the x-axis and the projection of the scattered vector on the xy-plane, spans from 0 degrees to 360 degrees. The code needs the function "pin_andrea" (eq. 4.47).